pyproject.toml
[project]
name = "almasim"
version = "2.1.10"
description = "An ALMA Simulation package for a more civilized era."
readme = "README.md"
requires-python = ">=3.12,<3.13"
authors = [
    {name = "Michele Delli Veneri", email = "micheledelliveneri@gmail.com"},
]
dependencies = [
    "astropy",
    "asyncssh",
    "pyvo",
    "numpy",
    "pandas",
    "matplotlib",
    "seaborn",
    "tqdm",
    "scipy",
    "h5py",
    "kaggle",
    "dask",
    "dask-expr",
    "dask-jobqueue",
    "distributed",
    "astromartini",
    "Hdecompose",
    "paramiko",
    "pysftp",
    "tenacity",
    "nifty8",
    "setuptools",
    "multiprocess",
    "scikit-image",
    "imagecodecs",
    "casatools",
    "casatasks",
]

[project.optional-dependencies]
ms-casacore = ["python-casacore"]

[project.urls]
Repository = "https://github.com/MicheleDelliVeneri/ALMASim"
Homepage = "https://github.com/MicheleDelliVeneri/ALMASim"
Documentation = "https://almasim.readthedocs.io/en/latest"

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic-settings",
    "ipykernel>=7.2.0",
]
test = [
    "pytest",
    "pytest-cov",
    "fastapi",
    "uvicorn[standard]",
    "httpx",
    "pydantic-settings",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/almasim"]

[tool.ruff]
line-length = 100
exclude = ["experimental", "illustris_python", "node_modules", ".venv", "build", "htmlcov", "frontend"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["E501"]

[tool.ruff.format]
docstring-code-format = true

[tool.uv]
package = true

[tool.uv.workspace]
members = ["backend"]
